feat. 각 달에만 조회될게있게 수정, 랭킹 테이블 수정

+++ bgm-agit-api/src/main/java/com/bgmagitapi/kml/rank/dto/response/RankGetResponse.java
@@ -27,37 +27,37 @@ public class RankGetResponse {
     private Double tobiRate;           // 토비%
     private Double tobiMinus3Rate;     // 토비시3%
     
-    // ===== 순위 카운트 =====
-    private Integer firstCount;        // 1
-    private Integer secondCount;       // 2
-    private Integer thirdCount;        // 3
-    private Integer fourthCount;       // 4
-    
+    // ===== 순위 카운트 (동장 0.5 / 반장 1 가중치) =====
+    private Double firstCount;         // 1
+    private Double secondCount;        // 2
+    private Double thirdCount;         // 3
+    private Double fourthCount;        // 4
+
     // ===== 기타 =====
     private Double avgRank;            // 순위% (평균순위)
-    private Integer totalCount;        // 국수 (총 판수)
+    private Double totalCount;         // 국수 (가중 합산)
     // 여기까지가 프론트가 쓰는거임
-    
+
     // ===== 내부 계산용 (필요하면 유지) =====
-    private Integer plusCount;
-    private Integer minus2Count;
-    private Integer plus3Count;
-    private Integer tobiCount;
-    private Integer tobiMinus3Count;
-    
+    private Double plusCount;
+    private Double minus2Count;
+    private Double plus3Count;
+    private Double tobiCount;
+    private Double tobiMinus3Count;
+
     public RankGetResponse(Long memberId,
                            String memberNickname,
                            Double recordSumPoint,
-                           Integer totalCount,
-                           Integer firstCount,
-                           Integer secondCount,
-                           Integer thirdCount,
-                           Integer fourthCount,
-                           Integer plusCount,
-                           Integer minus2Count,
-                           Integer plus3Count,
-                           Integer tobiCount,
-                           Integer tobiMinus3Count) {
+                           Double totalCount,
+                           Double firstCount,
+                           Double secondCount,
+                           Double thirdCount,
+                           Double fourthCount,
+                           Double plusCount,
+                           Double minus2Count,
+                           Double plus3Count,
+                           Double tobiCount,
+                           Double tobiMinus3Count) {
         
         this.memberId = memberId;
         this.memberNickname = memberNickname;
